Adopted logic for unauthorized user on new map for new geostory
