+++ web/client/epics/feedbackMask.js
@@ -182,8 +182,8 @@ const feedbackMaskPromptLogin = (action$, store) => // TODO: separate login requ
     action$.ofType(MAP_CONFIG_LOAD_ERROR, DASHBOARD_LOAD_ERROR, LOAD_GEOSTORY_ERROR, CONTEXT_LOAD_ERROR, CONTEXT_LOAD_ERROR_CONTEXTCREATOR)
         .filter((action) => action.error
             && action.error.status === 403
-            && (pathnameSelector(store.getState()).indexOf("new") === -1 // new map has different handling (see redirectUnauthorizedUserOnNewMap, TODO: uniform different behaviour)
-                || pathnameSelector(store.getState()).indexOf("newgeostory") >= 0)) // geostory can use this (that is the same of the dashboard)
+            && pathnameSelector(store.getState()).indexOf("new") === -1 // new map and geostory has different handling (see redirectUnauthorizedUserOnNewLoadError, TODO: uniform different behaviour)
+            && pathnameSelector(store.getState()).indexOf("newgeostory") === -1)
         .filter(() => !isLoggedIn(store.getState()) && !isSharedStory(store.getState()))
         .exhaustMap(
             () =>
@@ -195,6 +195,15 @@ const feedbackMaskPromptLogin = (action$, store) => // TODO: separate login requ
 
                     ).takeUntil(action$.ofType(LOGIN_SUCCESS, LOCATION_CHANGE))
         );
+
+const redirectUnauthorizedUserOnNewLoadError = (action$, { getState = () => {}}) =>
+    action$.ofType(MAP_CONFIG_LOAD_ERROR, LOAD_GEOSTORY_ERROR)
+        .filter((action) => action.error &&
+            action.error.status === 403 &&
+            (pathnameSelector(getState()).indexOf("new") !== -1 ||
+            pathnameSelector(getState()).indexOf("newgeostory") !== -1))
+        .filter(() => !isLoggedIn(getState()))
+        .switchMap(() => Rx.Observable.of(push('/'))); // go to home page
 
 /**
  * Epics for feedbackMask functionality
@@ -208,5 +217,6 @@ module.exports = {
     updateDashboardVisibility,
     updateGeoStoryFeedbackMaskVisibility,
     detectNewPage,
-    feedbackMaskPromptLogin
+    feedbackMaskPromptLogin,
+    redirectUnauthorizedUserOnNewLoadError
 };
